+++ src/Model/Export/Catalog/ProductType/BundleDataProvider.php
@@ -6,6 +6,7 @@ namespace Omikron\Factfinder\Model\Export\Catalog\ProductType;
 
 use Magento\Bundle\Model\Product\CatalogPrice;
 use Magento\Catalog\Model\Product;
+use Magento\CatalogInventory\Model\Stock\StockItemRepository as StockItem;
 use Magento\Directory\Model\PriceCurrency;
 use Omikron\Factfinder\Model\Formatter\NumberFormatter;
 
@@ -16,14 +17,16 @@ class BundleDataProvider extends SimpleDataProvider
         protected NumberFormatter $numberFormatter,
         private readonly PriceCurrency     $priceCurrency,
         private readonly CatalogPrice      $priceModel,
+        StockItem $stockItem,
         protected array             $productFields = []
     ) {
-        parent::__construct($product, $numberFormatter, $productFields);
+        parent::__construct($product, $numberFormatter, $stockItem, $productFields);
     }
 
     public function toArray(): array
     {
         $price = (float) $this->priceCurrency->convert($this->priceModel->getCatalogPrice($this->product));
+
         return ['Price' => $this->numberFormatter->format($price)] + parent::toArray();
     }
 }