+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 ## Unreleased
 ### Fix
 - Fix availability column data for UI export type
+- Fix deprecations from phpcs
+### Add
+- Add phpcs fixer as dev dependencies
+
 ## [v4.3.0] - 2024.01.16
 ### Add
 - Add landing page campaign