+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## Unreleased
+## [v4.3.1] - 2024.04.02
 ### Fix
 - Fix availability column data for UI export type
 - Fix deprecations from phpcs
@@ -601,6 +601,7 @@
 ### Added
 - Feed Export: Export feed file is now available via separate link
 
+[v4.3.1]:       https://github.com/FACT-Finder-Web-Components/magento2-module/releases/tag/v4.3.1
 [v4.3.0]:       https://github.com/FACT-Finder-Web-Components/magento2-module/releases/tag/v4.3.0
 [v4.2.2]:       https://github.com/FACT-Finder-Web-Components/magento2-module/releases/tag/v4.2.2
 [v4.2.1]:       https://github.com/FACT-Finder-Web-Components/magento2-module/releases/tag/v4.2.1